Migrate admin type to web type
