--- packages/project-builder-lib/src/schema/apps/admin/sections/crud-form/index.ts
@@ -1,4 +1,0 @@
-export * from './admin-crud-input.js';
-export * from './admin-input-spec.js';
-export * from './built-in-input.js';
-export * from './types.js';
