--- packages/project-builder-lib/src/schema/apps/admin/sections/index.ts
@@ -1,4 +1,0 @@
-export * from './base.js';
-export * from './crud-form/index.js';
-export * from './crud.js';
-export * from './types.js';
