--- packages/project-builder-lib/src/schema/apps/admin/sections/types.ts
@@ -1,7 +1,0 @@
-import { createEntityType } from '#src/references/index.js';
-
-import { appEntityType } from '../../types.js';
-
-export const adminSectionEntityType = createEntityType('admin-section', {
-  parentType: appEntityType,
-});
