+++ packages/project-builder-lib/src/schema/apps/index.ts
@@ -1,5 +1,7 @@
-export * from './admin/index.js';
 export * from './backend/index.js';
 export * from './base.js';
 export * from './types.js';
+// Re-export admin types from web admin for backward compatibility
+export * from './web/admin/index.js';
+
 export * from './web/index.js';
