+++ packages/project-builder-lib/src/schema/apps/web/web-app.ts
@@ -7,27 +7,7 @@ import { definitionSchema } from '#src/schema/creator/schema-creator.js';
 
 import { baseAppValidators } from '../base.js';
 import { createAppEntryType } from '../types.js';
-import { createAdminCrudSectionSchema } from './admin/sections/crud.js';
-import { adminSectionEntityType } from './admin/sections/types.js';
-
-// Admin section schema for web apps (copied from admin app schema)
-export const createWebAdminSectionSchema = definitionSchema((ctx) =>
-  ctx.withRefBuilder(createAdminCrudSectionSchema(ctx), (builder) => {
-    builder.addEntity({
-      type: adminSectionEntityType,
-      parentPath: { context: 'app' },
-      addContext: 'admin-section',
-    });
-  }),
-);
-
-export type WebAdminSectionConfig = def.InferOutput<
-  typeof createWebAdminSectionSchema
->;
-
-export type WebAdminSectionConfigInput = def.InferInput<
-  typeof createWebAdminSectionSchema
->;
+import { createAdminAppSchema } from './admin/admin.js';
 
 export const createWebAppSchema = definitionSchema((ctx) =>
   z.object({
@@ -47,26 +27,7 @@ export const createWebAppSchema = definitionSchema((ctx) =>
     ),
     includeUploadComponents: ctx.withDefault(z.boolean(), false),
     enableSubscriptions: ctx.withDefault(z.boolean(), false),
-    adminConfig: ctx.withDefault(
-      z.object({
-        enabled: z.boolean(),
-        pathPrefix: z.string().default('/admin'),
-        allowedRoles: ctx.withDefault(
-          z.array(
-            ctx.withRef({
-              type: authRoleEntityType,
-              onDelete: 'DELETE',
-            }),
-          ),
-          [],
-        ),
-        sections: ctx.withDefault(
-          z.array(createWebAdminSectionSchema(ctx)),
-          [],
-        ),
-      }),
-      { enabled: false, pathPrefix: '/admin' },
-    ),
+    adminApp: createAdminAppSchema(ctx),
   }),
 );
