+++ packages/project-builder-server/src/compiler/web/index.ts
@@ -3,24 +3,107 @@ import type {
   ProjectDefinitionContainer,
   WebAppConfig,
 } from '@baseplate-dev/project-builder-lib';
+import type { AdminLayoutLinkItem } from '@baseplate-dev/react-generators';
 import type { GeneratorBundle } from '@baseplate-dev/sync';
 
 import { composeNodeGenerator } from '@baseplate-dev/core-generators';
-import { AppUtils, webAppEntryType } from '@baseplate-dev/project-builder-lib';
 import {
+  AppUtils,
+  FeatureUtils,
+  webAppEntryType,
+} from '@baseplate-dev/project-builder-lib';
+import {
+  adminBullBoardGenerator,
+  adminComponentsGenerator,
+  adminHomeGenerator,
+  adminLayoutGenerator,
   apolloErrorGenerator,
   apolloErrorLinkGenerator,
   apolloSentryGenerator,
   composeReactGenerators,
   reactApolloGenerator,
+  reactComponentsGenerator,
   reactRouterGenerator,
+  reactRoutesGenerator,
   reactSentryGenerator,
   reactTailwindGenerator,
 } from '@baseplate-dev/react-generators';
 import { safeMerge } from '@baseplate-dev/utils';
 
+import { dasherizeCamel, titleizeCamel } from '#src/utils/case.js';
+
 import { AppEntryBuilder } from '../app-entry-builder.js';
 import { compileWebFeatures } from './features.js';
+
+function buildAdminNavigationLinks(
+  builder: AppEntryBuilder<WebAppConfig>,
+): AdminLayoutLinkItem[] {
+  const { adminApp } = builder.appConfig;
+  const { projectDefinition } = builder;
+
+  if (!adminApp?.enabled || !adminApp.sections) {
+    return [];
+  }
+
+  return adminApp.sections.map((section) => ({
+    type: 'link',
+    label: titleizeCamel(section.name),
+    icon: section.icon ?? 'MdHome',
+    path: `${adminApp.pathPrefix}/${
+      FeatureUtils.getFeatureByIdOrThrow(projectDefinition, section.featureRef)
+        .name
+    }/${dasherizeCamel(section.name)}`,
+  }));
+}
+
+function buildAdminRoutes(
+  builder: AppEntryBuilder<WebAppConfig>,
+): GeneratorBundle | undefined {
+  const { adminApp } = builder.appConfig;
+  const { projectDefinition } = builder;
+
+  if (!adminApp?.enabled) {
+    return undefined;
+  }
+
+  const backendApp = AppUtils.getBackendApp(projectDefinition);
+  const generalSettings = projectDefinition.settings.general;
+
+  return reactRoutesGenerator({
+    name: 'admin',
+    children: {
+      adminLayout: adminLayoutGenerator({
+        links: [
+          {
+            type: 'link',
+            label: 'Home',
+            icon: 'MdHome',
+            path: adminApp.pathPrefix,
+          },
+          ...buildAdminNavigationLinks(builder),
+          ...(backendApp.enableBullQueue
+            ? [
+                {
+                  type: 'link' as const,
+                  label: 'Queues',
+                  icon: 'AiOutlineOrderedList',
+                  path: `${adminApp.pathPrefix}/bull-board`,
+                },
+              ]
+            : []),
+        ],
+      }),
+      admin: adminHomeGenerator({}),
+      adminRoutes: backendApp.enableBullQueue
+        ? adminBullBoardGenerator({
+            bullBoardUrl: `http://localhost:${generalSettings.portOffset + 1}`,
+          })
+        : undefined,
+      // TODO: Implement admin sections compilation
+      // routes: compileAdminSections(builder),
+    },
+  });
+}
 
 function buildReact(builder: AppEntryBuilder<WebAppConfig>): GeneratorBundle {
   const { projectDefinition, appConfig, appCompiler } = builder;
@@ -32,6 +115,15 @@ function buildReact(builder: AppEntryBuilder<WebAppConfig>): GeneratorBundle {
   );
 
   const rootFeatures = appCompiler.getRootChildren();
+  const adminRoutes = buildAdminRoutes(builder);
+
+  const routerChildren = safeMerge(
+    {
+      features: compileWebFeatures(builder),
+      ...(adminRoutes ? { adminRoute: adminRoutes } : {}),
+    },
+    rootFeatures,
+  );
 
   return composeReactGenerators(
     {
@@ -39,14 +131,10 @@ function buildReact(builder: AppEntryBuilder<WebAppConfig>): GeneratorBundle {
       description: appConfig.description,
       children: {
         reactRouter: reactRouterGenerator({
-          children: safeMerge(
-            {
-              features: compileWebFeatures(builder),
-            },
-            rootFeatures,
-          ),
+          children: routerChildren,
           renderPlaceholderIndex: true,
         }),
+        reactComponents: reactComponentsGenerator({}),
         reactTailwind: reactTailwindGenerator({}),
         reactSentry: reactSentryGenerator({}),
         reactApollo: reactApolloGenerator({
@@ -59,6 +147,9 @@ function buildReact(builder: AppEntryBuilder<WebAppConfig>): GeneratorBundle {
           },
         }),
         apolloError: apolloErrorGenerator({}),
+        ...(appConfig.adminApp?.enabled
+          ? { adminComponents: adminComponentsGenerator({}) }
+          : {}),
       },
     },
     {
