+++ packages/project-builder-web/src/route-tree.gen.ts
@@ -32,16 +32,12 @@ import { Route as AppsEditKeyBackendRouteImport } from './routes/apps/edit.$key/
 import { Route as DataModelsEditKeyRouteRouteImport } from './routes/data/models/edit.$key/route'
 import { Route as DataEnumsEditKeyRouteRouteImport } from './routes/data/enums/edit.$key/route'
 import { Route as AppsEditKeyWebRouteRouteImport } from './routes/apps/edit.$key/web/route'
-import { Route as AppsEditKeyAdminRouteRouteImport } from './routes/apps/edit.$key/admin/route'
 import { Route as DataModelsEditKeyIndexRouteImport } from './routes/data/models/edit.$key/index'
 import { Route as DataEnumsEditKeyIndexRouteImport } from './routes/data/enums/edit.$key/index'
 import { Route as AppsEditKeyWebIndexRouteImport } from './routes/apps/edit.$key/web/index'
-import { Route as AppsEditKeyAdminIndexRouteImport } from './routes/apps/edit.$key/admin/index'
 import { Route as DataModelsEditKeyServiceRouteImport } from './routes/data/models/edit.$key/service'
 import { Route as DataModelsEditKeyGraphqlRouteImport } from './routes/data/models/edit.$key/graphql'
-import { Route as AppsEditKeyAdminSectionsRouteImport } from './routes/apps/edit.$key/admin/sections'
 import { Route as AppsEditKeyWebAdminIndexRouteImport } from './routes/apps/edit.$key/web/admin/index'
-import { Route as AppsEditKeyAdminSectionsSectionKeyRouteImport } from './routes/apps/edit.$key/admin/sections.$sectionKey'
 
 const SettingsRouteRoute = SettingsRouteRouteImport.update({
   id: '/settings',
@@ -159,11 +155,6 @@ const AppsEditKeyWebRouteRoute = AppsEditKeyWebRouteRouteImport.update({
   path: '/web',
   getParentRoute: () => AppsEditKeyRouteRoute,
 } as any)
-const AppsEditKeyAdminRouteRoute = AppsEditKeyAdminRouteRouteImport.update({
-  id: '/admin',
-  path: '/admin',
-  getParentRoute: () => AppsEditKeyRouteRoute,
-} as any)
 const DataModelsEditKeyIndexRoute = DataModelsEditKeyIndexRouteImport.update({
   id: '/',
   path: '/',
@@ -179,11 +170,6 @@ const AppsEditKeyWebIndexRoute = AppsEditKeyWebIndexRouteImport.update({
   path: '/',
   getParentRoute: () => AppsEditKeyWebRouteRoute,
 } as any)
-const AppsEditKeyAdminIndexRoute = AppsEditKeyAdminIndexRouteImport.update({
-  id: '/',
-  path: '/',
-  getParentRoute: () => AppsEditKeyAdminRouteRoute,
-} as any)
 const DataModelsEditKeyServiceRoute =
   DataModelsEditKeyServiceRouteImport.update({
     id: '/service',
@@ -196,23 +182,11 @@ const DataModelsEditKeyGraphqlRoute =
     path: '/graphql',
     getParentRoute: () => DataModelsEditKeyRouteRoute,
   } as any)
-const AppsEditKeyAdminSectionsRoute =
-  AppsEditKeyAdminSectionsRouteImport.update({
-    id: '/sections',
-    path: '/sections',
-    getParentRoute: () => AppsEditKeyAdminRouteRoute,
-  } as any)
 const AppsEditKeyWebAdminIndexRoute =
   AppsEditKeyWebAdminIndexRouteImport.update({
     id: '/admin/',
     path: '/admin/',
     getParentRoute: () => AppsEditKeyWebRouteRoute,
-  } as any)
-const AppsEditKeyAdminSectionsSectionKeyRoute =
-  AppsEditKeyAdminSectionsSectionKeyRouteImport.update({
-    id: '/$sectionKey',
-    path: '/$sectionKey',
-    getParentRoute: () => AppsEditKeyAdminSectionsRoute,
   } as any)
 
 export interface FileRoutesByFullPath {
@@ -234,20 +208,16 @@ export interface FileRoutesByFullPath {
   '/plugins/edit/$id': typeof PluginsEditIdRoute
   '/data/enums/': typeof DataEnumsIndexRoute
   '/data/models/': typeof DataModelsIndexRoute
-  '/apps/edit/$key/admin': typeof AppsEditKeyAdminRouteRouteWithChildren
   '/apps/edit/$key/web': typeof AppsEditKeyWebRouteRouteWithChildren
   '/data/enums/edit/$key': typeof DataEnumsEditKeyRouteRouteWithChildren
   '/data/models/edit/$key': typeof DataModelsEditKeyRouteRouteWithChildren
   '/apps/edit/$key/backend': typeof AppsEditKeyBackendRoute
   '/apps/edit/$key/': typeof AppsEditKeyIndexRoute
-  '/apps/edit/$key/admin/sections': typeof AppsEditKeyAdminSectionsRouteWithChildren
   '/data/models/edit/$key/graphql': typeof DataModelsEditKeyGraphqlRoute
   '/data/models/edit/$key/service': typeof DataModelsEditKeyServiceRoute
-  '/apps/edit/$key/admin/': typeof AppsEditKeyAdminIndexRoute
   '/apps/edit/$key/web/': typeof AppsEditKeyWebIndexRoute
   '/data/enums/edit/$key/': typeof DataEnumsEditKeyIndexRoute
   '/data/models/edit/$key/': typeof DataModelsEditKeyIndexRoute
-  '/apps/edit/$key/admin/sections/$sectionKey': typeof AppsEditKeyAdminSectionsSectionKeyRoute
   '/apps/edit/$key/web/admin': typeof AppsEditKeyWebAdminIndexRoute
 }
 export interface FileRoutesByTo {
@@ -264,14 +234,11 @@ export interface FileRoutesByTo {
   '/data/models': typeof DataModelsIndexRoute
   '/apps/edit/$key/backend': typeof AppsEditKeyBackendRoute
   '/apps/edit/$key': typeof AppsEditKeyIndexRoute
-  '/apps/edit/$key/admin/sections': typeof AppsEditKeyAdminSectionsRouteWithChildren
   '/data/models/edit/$key/graphql': typeof DataModelsEditKeyGraphqlRoute
   '/data/models/edit/$key/service': typeof DataModelsEditKeyServiceRoute
-  '/apps/edit/$key/admin': typeof AppsEditKeyAdminIndexRoute
   '/apps/edit/$key/web': typeof AppsEditKeyWebIndexRoute
   '/data/enums/edit/$key': typeof DataEnumsEditKeyIndexRoute
   '/data/models/edit/$key': typeof DataModelsEditKeyIndexRoute
-  '/apps/edit/$key/admin/sections/$sectionKey': typeof AppsEditKeyAdminSectionsSectionKeyRoute
   '/apps/edit/$key/web/admin': typeof AppsEditKeyWebAdminIndexRoute
 }
 export interface FileRoutesById {
@@ -294,20 +261,16 @@ export interface FileRoutesById {
   '/plugins/edit/$id': typeof PluginsEditIdRoute
   '/data/enums/': typeof DataEnumsIndexRoute
   '/data/models/': typeof DataModelsIndexRoute
-  '/apps/edit/$key/admin': typeof AppsEditKeyAdminRouteRouteWithChildren
   '/apps/edit/$key/web': typeof AppsEditKeyWebRouteRouteWithChildren
   '/data/enums/edit/$key': typeof DataEnumsEditKeyRouteRouteWithChildren
   '/data/models/edit/$key': typeof DataModelsEditKeyRouteRouteWithChildren
   '/apps/edit/$key/backend': typeof AppsEditKeyBackendRoute
   '/apps/edit/$key/': typeof AppsEditKeyIndexRoute
-  '/apps/edit/$key/admin/sections': typeof AppsEditKeyAdminSectionsRouteWithChildren
   '/data/models/edit/$key/graphql': typeof DataModelsEditKeyGraphqlRoute
   '/data/models/edit/$key/service': typeof DataModelsEditKeyServiceRoute
-  '/apps/edit/$key/admin/': typeof AppsEditKeyAdminIndexRoute
   '/apps/edit/$key/web/': typeof AppsEditKeyWebIndexRoute
   '/data/enums/edit/$key/': typeof DataEnumsEditKeyIndexRoute
   '/data/models/edit/$key/': typeof DataModelsEditKeyIndexRoute
-  '/apps/edit/$key/admin/sections/$sectionKey': typeof AppsEditKeyAdminSectionsSectionKeyRoute
   '/apps/edit/$key/web/admin/': typeof AppsEditKeyWebAdminIndexRoute
 }
 export interface FileRouteTypes {
@@ -331,20 +294,16 @@ export interface FileRouteTypes {
     | '/plugins/edit/$id'
     | '/data/enums/'
     | '/data/models/'
-    | '/apps/edit/$key/admin'
     | '/apps/edit/$key/web'
     | '/data/enums/edit/$key'
     | '/data/models/edit/$key'
     | '/apps/edit/$key/backend'
     | '/apps/edit/$key/'
-    | '/apps/edit/$key/admin/sections'
     | '/data/models/edit/$key/graphql'
     | '/data/models/edit/$key/service'
-    | '/apps/edit/$key/admin/'
     | '/apps/edit/$key/web/'
     | '/data/enums/edit/$key/'
     | '/data/models/edit/$key/'
-    | '/apps/edit/$key/admin/sections/$sectionKey'
     | '/apps/edit/$key/web/admin'
   fileRoutesByTo: FileRoutesByTo
   to:
@@ -361,14 +320,11 @@ export interface FileRouteTypes {
     | '/data/models'
     | '/apps/edit/$key/backend'
     | '/apps/edit/$key'
-    | '/apps/edit/$key/admin/sections'
     | '/data/models/edit/$key/graphql'
     | '/data/models/edit/$key/service'
-    | '/apps/edit/$key/admin'
     | '/apps/edit/$key/web'
     | '/data/enums/edit/$key'
     | '/data/models/edit/$key'
-    | '/apps/edit/$key/admin/sections/$sectionKey'
     | '/apps/edit/$key/web/admin'
   id:
     | '__root__'
@@ -390,20 +346,16 @@ export interface FileRouteTypes {
     | '/plugins/edit/$id'
     | '/data/enums/'
     | '/data/models/'
-    | '/apps/edit/$key/admin'
     | '/apps/edit/$key/web'
     | '/data/enums/edit/$key'
     | '/data/models/edit/$key'
     | '/apps/edit/$key/backend'
     | '/apps/edit/$key/'
-    | '/apps/edit/$key/admin/sections'
     | '/data/models/edit/$key/graphql'
     | '/data/models/edit/$key/service'
-    | '/apps/edit/$key/admin/'
     | '/apps/edit/$key/web/'
     | '/data/enums/edit/$key/'
     | '/data/models/edit/$key/'
-    | '/apps/edit/$key/admin/sections/$sectionKey'
     | '/apps/edit/$key/web/admin/'
   fileRoutesById: FileRoutesById
 }
@@ -578,13 +530,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AppsEditKeyWebRouteRouteImport
       parentRoute: typeof AppsEditKeyRouteRoute
     }
-    '/apps/edit/$key/admin': {
-      id: '/apps/edit/$key/admin'
-      path: '/admin'
-      fullPath: '/apps/edit/$key/admin'
-      preLoaderRoute: typeof AppsEditKeyAdminRouteRouteImport
-      parentRoute: typeof AppsEditKeyRouteRoute
-    }
     '/data/models/edit/$key/': {
       id: '/data/models/edit/$key/'
       path: '/'
@@ -606,13 +551,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AppsEditKeyWebIndexRouteImport
       parentRoute: typeof AppsEditKeyWebRouteRoute
     }
-    '/apps/edit/$key/admin/': {
-      id: '/apps/edit/$key/admin/'
-      path: '/'
-      fullPath: '/apps/edit/$key/admin/'
-      preLoaderRoute: typeof AppsEditKeyAdminIndexRouteImport
-      parentRoute: typeof AppsEditKeyAdminRouteRoute
-    }
     '/data/models/edit/$key/service': {
       id: '/data/models/edit/$key/service'
       path: '/service'
@@ -627,13 +565,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof DataModelsEditKeyGraphqlRouteImport
       parentRoute: typeof DataModelsEditKeyRouteRoute
     }
-    '/apps/edit/$key/admin/sections': {
-      id: '/apps/edit/$key/admin/sections'
-      path: '/sections'
-      fullPath: '/apps/edit/$key/admin/sections'
-      preLoaderRoute: typeof AppsEditKeyAdminSectionsRouteImport
-      parentRoute: typeof AppsEditKeyAdminRouteRoute
-    }
     '/apps/edit/$key/web/admin/': {
       id: '/apps/edit/$key/web/admin/'
       path: '/admin'
@@ -641,45 +572,8 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AppsEditKeyWebAdminIndexRouteImport
       parentRoute: typeof AppsEditKeyWebRouteRoute
     }
-    '/apps/edit/$key/admin/sections/$sectionKey': {
-      id: '/apps/edit/$key/admin/sections/$sectionKey'
-      path: '/$sectionKey'
-      fullPath: '/apps/edit/$key/admin/sections/$sectionKey'
-      preLoaderRoute: typeof AppsEditKeyAdminSectionsSectionKeyRouteImport
-      parentRoute: typeof AppsEditKeyAdminSectionsRoute
-    }
   }
 }
-
-interface AppsEditKeyAdminSectionsRouteChildren {
-  AppsEditKeyAdminSectionsSectionKeyRoute: typeof AppsEditKeyAdminSectionsSectionKeyRoute
-}
-
-const AppsEditKeyAdminSectionsRouteChildren: AppsEditKeyAdminSectionsRouteChildren =
-  {
-    AppsEditKeyAdminSectionsSectionKeyRoute:
-      AppsEditKeyAdminSectionsSectionKeyRoute,
-  }
-
-const AppsEditKeyAdminSectionsRouteWithChildren =
-  AppsEditKeyAdminSectionsRoute._addFileChildren(
-    AppsEditKeyAdminSectionsRouteChildren,
-  )
-
-interface AppsEditKeyAdminRouteRouteChildren {
-  AppsEditKeyAdminSectionsRoute: typeof AppsEditKeyAdminSectionsRouteWithChildren
-  AppsEditKeyAdminIndexRoute: typeof AppsEditKeyAdminIndexRoute
-}
-
-const AppsEditKeyAdminRouteRouteChildren: AppsEditKeyAdminRouteRouteChildren = {
-  AppsEditKeyAdminSectionsRoute: AppsEditKeyAdminSectionsRouteWithChildren,
-  AppsEditKeyAdminIndexRoute: AppsEditKeyAdminIndexRoute,
-}
-
-const AppsEditKeyAdminRouteRouteWithChildren =
-  AppsEditKeyAdminRouteRoute._addFileChildren(
-    AppsEditKeyAdminRouteRouteChildren,
-  )
 
 interface AppsEditKeyWebRouteRouteChildren {
   AppsEditKeyWebIndexRoute: typeof AppsEditKeyWebIndexRoute
@@ -695,14 +589,12 @@ const AppsEditKeyWebRouteRouteWithChildren =
   AppsEditKeyWebRouteRoute._addFileChildren(AppsEditKeyWebRouteRouteChildren)
 
 interface AppsEditKeyRouteRouteChildren {
-  AppsEditKeyAdminRouteRoute: typeof AppsEditKeyAdminRouteRouteWithChildren
   AppsEditKeyWebRouteRoute: typeof AppsEditKeyWebRouteRouteWithChildren
   AppsEditKeyBackendRoute: typeof AppsEditKeyBackendRoute
   AppsEditKeyIndexRoute: typeof AppsEditKeyIndexRoute
 }
 
 const AppsEditKeyRouteRouteChildren: AppsEditKeyRouteRouteChildren = {
-  AppsEditKeyAdminRouteRoute: AppsEditKeyAdminRouteRouteWithChildren,
   AppsEditKeyWebRouteRoute: AppsEditKeyWebRouteRouteWithChildren,
   AppsEditKeyBackendRoute: AppsEditKeyBackendRoute,
   AppsEditKeyIndexRoute: AppsEditKeyIndexRoute,
